# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HRM Platform API",
        "description": "API documentation for HRM Platform",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/performance/reviews": {
            "get": {
                "tags": [
                    "Performance"
                ],
                "summary": "Get performance reviews",
                "operationId": "071386bcef1cec60f16125864dad10fb",
                "responses": {
                    "200": {
                        "description": "Performance reviews list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Performance"
                ],
                "summary": "Create a performance review",
                "operationId": "d8853b05ede8739dded7f4c905cc5b24",
                "responses": {
                    "200": {
                        "description": "Performance review created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/performance/reviews/{id}": {
            "put": {
                "tags": [
                    "Performance"
                ],
                "summary": "Update a performance review",
                "operationId": "2495c0562e2b9f836455de35c869a707",
                "responses": {
                    "200": {
                        "description": "Performance review updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/performance/okrs": {
            "get": {
                "tags": [
                    "Performance"
                ],
                "summary": "Get OKRs",
                "operationId": "8a4888db43973a07337fe1f0de5ad257",
                "responses": {
                    "200": {
                        "description": "OKRs list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Performance"
                ],
                "summary": "Create an OKR",
                "operationId": "9a926dc32ba52fde0aba6d70380e30c9",
                "responses": {
                    "200": {
                        "description": "OKR created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/performance/okrs/{id}": {
            "put": {
                "tags": [
                    "Performance"
                ],
                "summary": "Update an OKR",
                "operationId": "290570be65cb25cd12c866fe1d5ce037",
                "responses": {
                    "200": {
                        "description": "OKR updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/performance/goals": {
            "get": {
                "tags": [
                    "Performance"
                ],
                "summary": "Get goals",
                "operationId": "1e97577826e356ff0063af09029a1f5a",
                "responses": {
                    "200": {
                        "description": "Goals list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Performance"
                ],
                "summary": "Create a goal",
                "operationId": "923b0840a87560ae02e20f020d5bb4aa",
                "responses": {
                    "200": {
                        "description": "Goal created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/performance/goals/{id}": {
            "put": {
                "tags": [
                    "Performance"
                ],
                "summary": "Update a goal",
                "operationId": "83d6d3e95aebcf39bec84d244c9ef4e3",
                "responses": {
                    "200": {
                        "description": "Goal updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/performance/summary": {
            "get": {
                "tags": [
                    "Performance"
                ],
                "summary": "Get performance summary for dashboard",
                "operationId": "99a447c714081dd79682af6fc3ba5501",
                "responses": {
                    "200": {
                        "description": "Performance summary"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}